+++ src/app/config/svg.d.ts
