+++ src/app/store/atoms.ts
@@ -10,7 +10,6 @@ export const isLoggedInState = atom({
 export const userInfoState = atom({
   key: 'userInfoState',
   default: { _id: '', displayName: '', phoneNumber: '', email: '' },
-  // default: {},
 });
 export const searchContentsState = atom({
   key: 'searchContentsState',
