+++ src/components/Footer.tsx
@@ -1,4 +1,4 @@
-import ResponsiveProvider from './WrapProvider/ResponsiveProvider';
+import ResponsiveProvider from './wrap-provider/ResponsiveProvider';
 import useDarkMode from '../shared/hooks/common/useDarkMode';
 export default function Footer() {
   const { isDarkMode } = useDarkMode();
