+++ src/components/Join/JoinForm.tsx
@@ -1,5 +1,5 @@
-import Button from '../AtomComponent/Button';
-import Input from '../AtomComponent/Input';
+import Button from '../atom-component/Button';
+import Input from '../atom-component/Input';
 import useJoin from '../../shared/hooks/auth/useJoin';
 import {
   passwordRegTest,
