+++ src/components/Login/LoginForm.tsx
@@ -1,6 +1,6 @@
 import { KeyboardEvent } from 'react';
-import Button from '../AtomComponent/Button';
-import Input from '../AtomComponent/Input';
+import Button from '../atom-component/Button';
+import Input from '../atom-component/Input';
 import { useNavigate } from 'react-router-dom';
 import { passwordRegTest, emailRegTest } from '../../shared/util/regs';
 import useLogin from '../../shared/hooks/auth/useLogin';
