+++ src/components/content-detail/ContentInfo.tsx
