+++ src/pages/Home.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import Banner from '../components/Banner/Banner';
+import Banner from '../components/banner/Banner';
 
 import CategoryContents from '../components/CategoryContents';
 import Footer from '../components/Footer';
