+++ src/pages/Login.tsx
@@ -1,6 +1,6 @@
-import ResponsiveProvider from '../components/WrapProvider/ResponsiveProvider';
+import ResponsiveProvider from '../components/wrap-provider/ResponsiveProvider';
 import FormHeader from '../components/FormHeader';
-import LoginForm from '../components/Login/LoginForm';
+import LoginForm from '../components/login/LoginForm';
 export default function Login() {
   return (
     <ResponsiveProvider direction={'col'}>
