+++ src/pages/ReviewDetail.tsx
@@ -7,8 +7,8 @@ import { formatDateWithTime } from '../shared/util/regs';
 import CategoryReviewList from '../components/CategoryContents';
 import Footer from '../components/Footer';
 import DetailBackground from '../components/DetailBackground';
-import Button from '../components/AtomComponent/Button';
-import ResponsiveProvider from '../components/WrapProvider/ResponsiveProvider';
+import Button from '../components/atom-component/Button';
+import ResponsiveProvider from '../components/wrap-provider/ResponsiveProvider';
 import useReviewDelete from '../shared/hooks/review/useReviewDelete';
 import useReviewSelect from '../shared/hooks/review/useReviewSelect';
 import Loading from '../components/Loading';
